Fixed Eigenvalue bug and implemented Z-Gate
By normalizing the value of psi rather than beta (used to derive psi), the Eigenvalue bug finally fixed. Unfortunately, certain phase shifts are still incorrectly represented on Bloch sphere under certain transformations.
TODO: experiment with stacked phase shifts in simulator, derive theta and psi states by hand, and compare to Bloch transformations in order to ensure that the simulator is accurate in it's rotations.  Likely cause of inaccuracy is initialization of the quantum circuit in |1> state.